Move YIELD_TOKEN list to utils file
