+++ containers/Balancer.ts
@@ -6,6 +6,7 @@ import { TOKENS, POOL } from "../apollo/balancer/queries";
 
 import Connection from "./Connection";
 import Token from "./Token";
+import { YIELD_TOKENS } from "../utils/yieldTokenList";
 
 interface PoolState {
   exitsCount: number;
@@ -37,35 +38,6 @@ interface yieldPair {
   [key: string]: string;
 }
 
-interface yieldToken {
-  [key: string]: yieldPair;
-}
-
-// The keys in this object are the synthetic token. the `token0` and `token1` are
-// the balancer pool key value pairs for the first and second token in the pool.
-const YIELD_TOKENS: yieldToken = {
-  "0x81ab848898b5ffd3354dbbefb333d5d183eedcb5": {
-    token0: "0x81ab848898b5ffd3354dbbefb333d5d183eedcb5",
-    token1: "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48",
-  }, // yUSDETH-SEP20
-  "0xb2fdd60ad80ca7ba89b9bab3b5336c2601c020b4": {
-    token0: "0xb2fdd60ad80ca7ba89b9bab3b5336c2601c020b4",
-    token1: "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48",
-  }, // yUSDETH-Oct20
-  "0x208d174775dc39fe18b1b374972f77ddec6c0f73": {
-    token0: "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48",
-    token1: "0x208d174775dc39fe18b1b374972f77ddec6c0f73",
-  }, // uUSDrBTC-OCT
-  "0xd16c79c8a39d44b2f3eb45d2019cd6a42b03e2a9": {
-    token0: "0xd16c79c8a39d44b2f3eb45d2019cd6a42b03e2a9",
-    token1: "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48",
-  }, // uUSDwETH-DEC
-  "0xf06ddacf71e2992e2122a1a0168c6967afdf63ce": {
-    token0: "0xf06ddacf71e2992e2122a1a0168c6967afdf63ce",
-    token1: "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48",
-  }, // uUSDrBTC-DEC
-};
-
 const useBalancer = () => {
   const { address, block$ } = Connection.useContainer();
   const { address: tokenAddress } = Token.useContainer();
@@ -82,7 +54,6 @@ const useBalancer = () => {
     string | null
   >(null);
   const [poolTokenList, setPoolTokenList] = useState<string[] | null>(null);
-  const [isYieldToken, setIsYieldToken] = useState<boolean>(false);
 
   const [usdPrice, setUsdPrice] = useState<number | null>(null);
   const [poolAddress, setPoolAddress] = useState<string | null>(null);
@@ -121,7 +92,6 @@ const useBalancer = () => {
       const IS_YIELD_TOKEN = Object.keys(YIELD_TOKENS).includes(
         tokenAddress.toLowerCase()
       );
-      setIsYieldToken(IS_YIELD_TOKEN);
       if (IS_YIELD_TOKEN) {
         setSelectedTokenAddress(tokenAddress.toLowerCase());
         setPoolTokenList([
@@ -338,8 +308,7 @@ const useBalancer = () => {
     usdPrice,
     shares,
     userShareFraction,
-    isYieldToken,
-    YIELD_TOKENS,
+    initializeTokenAddress,
     getPoolDataForToken,
     getTokenPrice,
   };
